CoRINs/server_corins/requirements-top-level.txt
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
matplotlib==3.2.2
pandas==1.0.5
